+++ src/Ordering/Ordering-Core/Domain/Entities/OrderItem.cs
@@ -18,5 +18,5 @@ public class OrderItem
     public Guid ProductId { get; set; } = default!;
     public string ProductName { get; set; } = default!;
     public int Quantity { get; set; } = default!;
-    public int Price { get; set; } = default!;
+    public long Price { get; set; } = default!;
 }
